Use angular velocity to update angle over time
Added a function get_angular_acceleration in utilities.py.

This function computes and returns the tangential component of gravity.

Added a private variable _angular_velocity to pendulum.py.

Added a private function _angular_velocity_change to pendulum.py.

This function adds angular acceleration to _angular_velocity.

The angular velocity is then added to the current angle over time.

Added a parameter named type to the pendulum's update function.

This allows you to pick which type of function to update the angle.